Move BundleSourceSection to opengever.bundle
