+++ opengever/bundle/cfgs/oggbundle.cfg
@@ -30,7 +30,7 @@ pipeline =
 
 
 [bundlesource]
-blueprint = opengever.setup.bundlesource
+blueprint = opengever.bundle.bundlesource
 
 [resolveguid]
 blueprint = opengever.setup.resolveguid
