+++ opengever/bundle/console.py
@@ -1,7 +1,7 @@
 from collective.transmogrifier.transmogrifier import Transmogrifier
+from opengever.bundle.sections.bundlesource import BUNDLE_PATH_KEY
 from opengever.core.debughelpers import get_first_plone_site
 from opengever.core.debughelpers import setup_plone
-from opengever.setup.sections.bundlesource import BUNDLE_PATH_KEY
 from zope.annotation import IAnnotations
 import logging
 import sys
@@ -26,7 +26,7 @@ def import_oggbundle(app, args):
     plone = setup_plone(get_first_plone_site(app))
     transmogrifier = Transmogrifier(plone)
     IAnnotations(transmogrifier)[BUNDLE_PATH_KEY] = bundle_path
-    transmogrifier(u'opengever.setup.oggbundle')
+    transmogrifier(u'opengever.bundle.oggbundle')
 
     log.info("Committing transaction...")
     transaction.commit()
