+++ opengever/bundle/transmogrifier.zcml
@@ -14,6 +14,11 @@
       />
 
   <utility
+      component=".sections.bundlesource.BundleSourceSection"
+      name="opengever.bundle.bundlesource"
+      />
+
+  <utility
       component=".sections.map_local_roles.MapLocalRolesSection"
       name="opengever.bundle.map_local_roles"
       />
